feat(experiments): 메인 개편을 main_redesign 가드레일 실험 뒤에 둔다
개편은 동아리 목록을 한 탭 뒤로 민다. 프리뷰에서 못 찾은 28.6%가 전체보기를
한 번 더 누를지 이탈할지는 관측 데이터에 없다. 그런 화면이 존재한 적이 없기
때문이다. 그래서 실험으로만 알 수 있다.

기존 메인은 LegacyMain으로 분리해 두 곳에서 쓴다. 태블릿·웹은 개편 대상이
아니라 항상 이 화면이고, 모바일·앱에서는 control이다. 노출 이벤트도 화면이
실제로 갈리는 모바일·앱에서만 보낸다.

7일 창·팔당 약 270명이라 잡히는 건 메인->동아리 상세 퍼널(baseline 89.8%)의
8%p 이상 하락뿐이다. 개선을 증명하는 실험이 아니라 크게 망가지지 않았는지만
보는 가드레일이다.

1차 지표를 카드 클릭의 page 속성으로 거르면 안 된다. 개편은 목록을 /clubs로
빼므로 treatment 클릭이 page=club-list로 찍혀, 행동이 나빠지지 않아도 지표만
폭락한다.
