+++ frontend/src/constants/eventName.ts
@@ -69,6 +69,10 @@ export const USER_EVENT = {
   // 하단 네비게이션
   BOTTOM_TAB_CLICKED: 'BottomTab Clicked',
 
+  HOME_SECTION_MORE_CLICKED: 'Home Section More Clicked',
+  HOME_SEARCH_CLICKED: 'Home Search Clicked',
+  HOME_SUBSCRIPTION_CLICKED: 'Home Subscription Clicked',
+
   // 동소한 (동아리 소개 한마당)
   FESTIVAL_TAB_CLICKED: 'Festival Tab Clicked',
   FESTIVAL_BOOTH_MAP_SLIDE_CHANGED: 'Festival BoothMap Slide Changed',
@@ -171,6 +175,7 @@ export const PAGE_VIEW = {
   APPLICATION_FORM_PAGE: 'ApplicationFormPage',
   CLUB_DETAIL_PAGE: 'ClubDetailPage',
   MAIN_PAGE: 'MainPage',
+  CLUB_LIST_PAGE: 'ClubListPage',
   SUBSCRIPTIONS_PAGE: 'SubscriptionsPage',
   MENU_PAGE: 'MenuPage',
   INTRODUCE_PAGE: 'IntroducePage',
@@ -204,6 +209,7 @@ export const PAGE_VIEW = {
 
 export const PAGE_NAME = {
   MAIN: 'main',
+  CLUB_LIST: 'club-list',
   WEBVIEW_MAIN: 'webview-main',
   INTRODUCE: 'introduce',
   SUBSCRIPTIONS: 'subscriptions',
