+++ frontend/src/experiments/CLAUDE.md
@@ -20,8 +20,8 @@ const variant = useExperimentVariant(mainBannerExperiment);
 **부팅 때 한꺼번에 보내면 안 된다.**
 
 배정은 사이트에 들어온 **모든 사람**에게 일어난다. 하지만 실험으로 화면이
-갈리는 건 그중 일부다. 예를 들어 개편이 모바일에만 적용되면 데스크톱
-사용자는 control이든 treatment든 **똑같은 화면**을 본다.
+갈리는 건 그중 일부다. `main_redesign`이 그렇다. 모바일·앱에서만 화면이
+갈리고 태블릿·웹 방문자는 control이든 treatment든 **똑같은 화면**을 본다.
 
 부팅할 때 노출 이벤트를 다 보내면 이 사람들도 실험 참가자로 집계된다. 같은
 화면을 보니 두 그룹에 절반씩 똑같이 섞이고, 결과적으로 진짜 차이를 묽게
@@ -32,12 +32,64 @@ const variant = useExperimentVariant(mainBannerExperiment);
 
 ```typescript
 useEffect(() => {
-  if (isNarrow) trackExperimentExposure(someExperiment);
+  if (isNarrow) trackExperimentExposure(mainRedesignExperiment);
 }, [isNarrow]);
 ```
 
 super property 등록(부팅 시)과 노출 이벤트(렌더 시)는 목적이 다르다. 전자는
 "이 방문의 배정이 무엇인가", 후자는 "누구를 분석 모집단에 넣을 것인가"다.
+
+## 진행 중인 실험
+
+| key | 변형 | 배분 | 목적 |
+| --- | --- | --- | --- |
+| `main_redesign` | control / treatment | 50:50 | 메인 개편 **가드레일** |
+
+`main_redesign`은 **7일 창에서는** 개선을 증명하는 실험이 아니다. 앱 메인 7일
+유니크 방문자가 535명이라 팔당 약 270명이고, 이 표본으로 잡히는 건 **8%p 이상의
+하락**뿐이다(검정력 76%). 5%p는 41%밖에 안 된다.
+
+- **1차 지표: 퍼널 `WebviewMainPage Visited` → `ClubDetailPage Visited`**
+  - baseline 89.8% (일자별 평균, 표준편차 3.1%p, 30일)
+- 중단 조건: 전환율 **8%p 이상** 하락
+- 2차 지표: 홍보 목록 유입률
+- 7일 안에 유의차 없음 = "개선했다"가 아니라 **"크게 망가지지 않았다"**로만 읽는다
+
+### 관측 창은 7일로 고정된 게 아니다
+
+아래 ITP 제약은 **웹 브라우저 얘기지 앱에는 그대로 걸리지 않는다.** 그리고 1차
+지표를 앱 퍼널로 정의했으므로 분석 모집단이 사실상 앱 유저다(메인 방문 30일
+기준 앱 1,124 / 웹 528).
+
+그래서 배포 후 **variant 뒤집힘 비율을 실측**해서 낮으면 창을 18일로 늘린다.
+18일이면 5%p를 검정력 80%로 잡아 가드레일이 아니라 실제 효과를 볼 수 있다.
+
+뒤집힘 실측은 지표가 아니라 데이터 품질 점검이므로 중간에 봐도 peeking이 아니다.
+반대로 **1차 지표는 창이 끝날 때까지 보지 않는다.**
+
+### 1차 지표를 `page`로 거르면 안 된다
+
+`ClubCard Clicked`를 `page = webview-main`으로 걸어 세면 **treatment가 구조적으로
+불리해진다.** 개편 홈은 프리뷰만 두고 목록을 `/clubs`로 뺐으므로 treatment의
+카드 클릭은 대부분 `page = club-list`로 찍힌다. 행동이 전혀 나빠지지 않아도
+지표만 폭락한다.
+
+그래서 "어디서 눌렀는지"를 묻지 않는 퍼널로 정의한다. 같은 이유로 카드 클릭
+계열 이벤트는 1차 지표로 쓰지 않는다.
+
+### 결과는 절대값이 아니라 두 팔의 차이로 읽는다
+
+전환율에 추세가 있다. 아무 변경이 없던 30일 안에서도 앞 7일 91.1% → 최근 7일
+86.0%로 5.1%p 내려갔다. 동시 배정이라 이 추세는 양팔에 똑같이 걸려 상쇄되지만,
+**treatment의 절대값만 보고 판단하면 추세를 개편 탓으로 오인한다.**
+
+하단바는 실험 밖이라 두 팔 모두 4탭을 본다. control은 "개편 전 프로덕션 전체"가
+아니라 **"하단바는 새것, 홈 화면만 기존"**이다. 그래서 기존 메인의 필터칩
+(동아리·홍보)은 control에서도 제거했다. 하단바 탭과 목적지가 같아 중복이고,
+남겨두면 홍보 진입점이 control만 둘이라 홍보 지표 비교가 깨진다.
+
+덕분에 두 팔의 차이는 **홈이 목록이냐 허브냐** 하나로 좁혀졌고, 홍보 목록
+유입률을 2차 지표로 볼 수 있다(진입점이 양쪽 다 하단바 탭 하나).
 
 ## 오염 계측
 
@@ -73,3 +125,22 @@ Mixpanel `distinct_id`가 배정과 함께 사라져, 재방문인지 신규 방
 - 배정부터 결과 지표까지의 관측 창을 7일 미만으로 잡아 ITP 만료 영향을 줄인다.
 - 로그인 사용자 대상이면 `hash(userId + key)` 기반 결정론적 버킷팅으로 바꾼다.
   localStorage에 의존하지 않으므로 위 경로가 전부 사라진다.
+
+### 단, 위 내용은 웹 브라우저에 한한다
+
+**앱 웹뷰는 사정이 다르다.** 메인 트래픽의 약 68%가 여기다.
+
+- 웹뷰가 `domStorageEnabled`에 `incognito` 미설정이라 **영구 저장소**를 쓴다.
+  앱을 껐다 켜도 배정이 유지되고, 사용자가 브라우저에서 사이트 데이터를 지우는
+  경로 자체가 없다. 앱 삭제·재설치나 기기 교체만 재추첨이다.
+- **식별자가 웹뷰 밖에 있다.** 앱이 `AsyncStorage`에 보관하는 설치 단위 ID를
+  URL로 넘겨주고 웹이 `mixpanel.identify()`로 쓴다(RN `utils/mixpanel.ts`의
+  `getOrCreateMixpanelSessionId`). 이름은 session이지만 실제로는 영구 ID다.
+
+두 저장소가 분리돼 있어서, 웹뷰 저장소만 날아가도 **같은 유저로 식별된 채
+variant만 바뀐다.** 즉 앱에서는 배정 뒤집힘이 **관측 가능하다** — 같은
+`distinct_id`에 실험 key 값이 둘 붙은 유저를 세면 된다. 위의 "측정할 수 없다"는
+웹 브라우저(식별자까지 같이 사라지는 경우)에만 해당한다.
+
+WKWebView 내부의 ITP 정책까지 코드로 확인한 것은 아니다. 그러므로 안전하다고
+단정하지 말고 **배포 후 위 방법으로 실측한 뒤** 관측 창을 늘릴지 정한다.
